pyproject.toml
[project]
name = "spacecraft-anomaly"
version = "0.1.0"
requires-python = ">=3.10"
dependencies = [
  "numpy",
  "pandas",
  "scikit-learn",
  "pyyaml",
  "scipy",
  "matplotlib",
  "torch",
  "torchmetrics",
  "tqdm",
]
